Function fix_message::parse

pub fn parse(inbound_message: &str) -> FIXMessageResult<FIXMessage>

This function validates and parses FIX message

Errors

Returns ChecksumFieldNotFound when 10 - ChekSum field not found in message:

extern crate fix_checksum;

use fix_message::*;
use fix_checksum::FIXChecksumValidatorError::*;

let mut message_parts = vec!["8=FIX.4.2", "9=73", "35=0", "49=BRKR",
  "56=INVMGR", "34=235", "52=19980604-07:58:28", "112=19980604-07:58:28"];
let mut message = message_parts.join(&(FIX_MESSAGE_DELIMITER.to_string()));

assert_eq!(parse(&message).unwrap_err(), InvalidChecksum(ChecksumFieldNotFound));

Returns ChecksumFieldInvalidFormat when 10 - ChekSum field value in invalid format:

extern crate fix_checksum;

use fix_message::*;
use fix_checksum::FIXChecksumValidatorError::*;

let message_parts = vec!["8=FIX.4.2", "9=73", "35=0", "49=BRKR", "56=INVMGR",
  "34=235", "52=19980604-07:58:28", "112=19980604-07:58:28", "10=2ZZ"];
let message = message_parts.join(&(FIX_MESSAGE_DELIMITER.to_string()));

assert_eq!(parse(&message).unwrap_err(),
  InvalidChecksum(ChecksumFieldInvalidFormat("2ZZ".parse::<u32>().unwrap_err())));

Returns InvalidChecksumValue when 10 - ChekSum field value is not valid:

use fix_message::*;

let message_parts = vec!["8=FIX.4.2", "9=73", "35=0", "49=BRKR", "56=INVMGR",
  "34=235", "52=19980604-07:58:28", "112=19980604-07:58:28", "10=231"];
let message = message_parts.join(&(FIX_MESSAGE_DELIMITER.to_string()));

assert_eq!(parse(&message).unwrap_err(), InvalidChecksumValue);

Returns InvalidFieldStructure when FIX message field does not have value:

use fix_message::*;

let message_parts = vec!["8=FIX.4.2", "9=73", "35=", "49=BRKR", "56=INVMGR",
  "34=235", "52=19980604-07:58:28", "112=19980604-07:58:28", "10=188"];
let message = message_parts.join(&(FIX_MESSAGE_DELIMITER.to_string()));

assert_eq!(parse(&message).unwrap_err(), InvalidFieldStructure);

Returns InvalidFirstField when the first field in message is not 8 - BeginString:

use fix_message::*;

let message_parts = vec!["9=FIX.4.2", "8=73", "35=0", "49=BRKR", "56=INVMGR",
  "34=235", "52=19980604-07:58:28", "112=19980604-07:58:28", "10=236"];
let message = message_parts.join(&(FIX_MESSAGE_DELIMITER.to_string()));

assert_eq!(parse(&message).unwrap_err(), InvalidFirstField("9"));

Returns InvalidSecondField when the second field in message is not 9 - BodyLength:

use fix_message::*;

let message_parts = vec!["8=FIX.4.2", "35=73", "9=0", "49=BRKR", "56=INVMGR",
  "34=235", "52=19980604-07:58:28", "112=19980604-07:58:28", "10=236"];
let message = message_parts.join(&(FIX_MESSAGE_DELIMITER.to_string()));

assert_eq!(parse(&message).unwrap_err(), InvalidSecondField("35"));

Returns InvalidThirdField when the third field in message is not 35 - MsgType:

use fix_message::*;

let message_parts = vec!["8=FIX.4.2", "9=73", "34=0", "49=BRKR", "56=INVMGR",
  "35=235", "52=19980604-07:58:28", "112=19980604-07:58:28", "10=236"];
let message = message_parts.join(&(FIX_MESSAGE_DELIMITER.to_string()));

assert_eq!(parse(&message).unwrap_err(), InvalidThirdField("34"));

Returns NotAllRequiredFieldsFound when not all fields required for standard header were found:

use fix_message::*;

let message_parts = vec!["8=FIX.4.2", "9=73", "35=0", "49=BRKR", "56=INVMGR",
  "52=19980604-07:58:28", "112=19980604-07:58:28", "10=173"];
let message = message_parts.join(&(FIX_MESSAGE_DELIMITER.to_string()));

assert_eq!(parse(&message).unwrap_err(), NotAllRequiredFieldsFound);

Returns ExtraRequiredFieldFound when additional required field has been found in standard header:

use fix_message::*;

let message_parts = vec!["8=FIX.4.2", "9=73", "35=0", "35=0", "49=BRKR", "56=INVMGR",
  "52=19980604-07:58:28", "112=19980604-07:58:28", "10=131"];
let message = message_parts.join(&(FIX_MESSAGE_DELIMITER.to_string()));

assert_eq!(parse(&message).unwrap_err(), ExtraRequiredFieldFound);

Examples

Basic usage:

use fix_message::*;

let message_parts = vec!["8=FIX.4.2", "9=73", "35=0", "49=BRKR", "56=INVMGR",
  "34=235", "52=19980604-07:58:28", "112=19980604-07:58:28", "10=236"];
let message = message_parts.join(&(FIX_MESSAGE_DELIMITER.to_string()));

let expected_fix_message = FIXMessage {
  version: "FIX.4.2".to_string(),
  data: vec![
    FIXMessageField { tag: "8".to_string(), value: "FIX.4.2".to_string() } ,
    FIXMessageField { tag: "9".to_string(), value: "73".to_string() } ,
    FIXMessageField { tag: "35".to_string(), value: "0".to_string() } ,
    FIXMessageField { tag: "49".to_string(), value: "BRKR".to_string() } ,
    FIXMessageField { tag: "56".to_string(), value: "INVMGR".to_string() } ,
    FIXMessageField { tag: "34".to_string(), value: "235".to_string() } ,
    FIXMessageField { tag: "52".to_string(), value: "19980604-07:58:28".to_string() } ,
    FIXMessageField { tag: "112".to_string(), value: "19980604-07:58:28".to_string() } ,
    FIXMessageField { tag: "10".to_string(), value: "236".to_string() }
  ]
};

 assert_eq!(parse(&message).unwrap(), expected_fix_message);